backend/requirements.txt
# --- Core FastAPI ---
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
sse-starlette>=1.8.0

# --- OpenAI ---
openai>=1.3.0

# --- Microsoft Agent Framework (core only — app uses ChatAgent, Runtime, openai) ---
agent-framework-core==1.0.0b251211

# --- Database (Turso) ---
libsql>=0.1.11

# --- Cache (Upstash Redis) ---
upstash-redis>=1.0.0

# --- Security & Auth ---
slowapi>=0.1.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
redis>=5.0.0  # For Upstash Redis TCP endpoint (SlowAPI compatible)

# --- AI Safety & Guardrails ---
guardrails-ai>=0.5.0

# --- PDF Processing ---
pymupdf4llm>=0.2.6
pymupdf>=1.23.0

# --- Utils ---
tqdm
python-dotenv
pydantic>=2.0.0
pydantic-settings>=2.0.0
structlog
rich
email-validator
numpy>=1.26.0

# --- HTTP & Web Scraping ---
httpx[http2]>=0.27.0
beautifulsoup4>=4.12.3

# --- LLM Providers ---
groq>=0.5.0

# --- Vector DB ---
upstash-vector>=0.3.0

# --- Additional PDF ---
pypdf>=4.0.0

# --- Payments (Stripe) ---
stripe>=8.0.0

# --- PDF Report Generation ---
reportlab>=4.0.0

# --- Email (Resend) ---
resend>=0.7.0

# --- Fiscal Calendar ---
icalendar>=6.0.0
pywebpush>=2.0.0

# --- Document Extraction (multi-format) ---
python-docx>=1.1.2
openpyxl>=3.1.0

# --- MFA/2FA (TOTP) ---
pyotp>=2.9.0
qrcode[pil]>=7.0

# --- Google SSO ---
google-auth>=2.0.0

# --- Google Gemini (Invoice OCR — Phase 3) ---
google-genai>=1.0.0
# cache-bust 1774479781

# --- Doc crawler (Scrapling + Playwright for AEAT/BOE/CCAA scraping) ---
scrapling==0.4.6
curl_cffi==0.15.0
playwright==1.58.0
browserforge==1.2.4

# DefensIA — generacion y parseo de documentos
lxml>=5.3.0
Jinja2>=3.1.4

# DefensIA — storage cifrado (Wave 2B T2B-013a)
cryptography>=42.0
zstandard>=0.22
